requirements.txt
pandas==2.2.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1